requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
tortoise-orm==0.17.8
aerich==0.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.7
bcrypt==4.0.1
gunicorn==21.2.0
asyncpg==0.28.0
psycopg2-binary==2.9.9
httpx==0.24.0
flake8==6.0.0
mypy==1.2.0
